Added message formatter override
There is now a Trace.FormatMessage property that takes either a null or a simple String Formatter(string message) delegate.

If a delegate is specified, it is called on evey message output by either Trace or Debug, and it can modify thos messages as it sees fit.

I use this to add time stamps to all messages, and to colorize messages by source.